added ft_unsetenv + tests
unsetenv added and tests.
unsetenv removes the key and value for the given key from the env.
If the first element is to be deleted,
it copies the values of the second element to the first,
and then deletes the second.
If there is no second element, the key and value are emptied,
so the element can later be filled.