Hotfix/freeenvlistcontent (#20)
* fixed compile error in memccpy

was dereferencing a void pointer, so gcc complained. Fixed that.

* added more explicit checking in memexpand

to avoid erronous or unneeded malloc calls,
I added a condition to the second block here.

* fixed leaks on ctrl+d exit

when the user presses ctrl+d, everything is now correctly freed.

* solved huge leaks in free_env_list

I was not freeing the content of the list, just the structs themself.

* converted the exit builtin to be a void function.

Since this calls exit, returing never happens.

* fixed a indexing mistake in GNL

here I was supposed to check if the last character of
the read buffer was a newline, but I got the index wrong.
This caused a static list element to be created when not needed.