Hotfix/tabsandexpansionincommand (#26)
* fixed expansions not working in argv0

I assumed expansions in the command itself were not valid.
Fixed that, and made error when execution fails a bit clearer.

* added strsplit_t_s but has VERY weird behaviour.

* fixed stupid mistake in strsplit_t_s

fixed the breaking mistake I have been reading over for the past HOURS..
Also cleaned up the mallocs.

* Implemented argument seperation by tabs

arguments can now also be seperated by tabs,
thanks to a new implementation of ft_strsplit. (hardcoded to tabs and spaces)
This is to properly and reliably handle arguments.