fix: fixed infinite loop in variable expansion
when FOO=$FOO is in the env,
it would create an infinite loop when expanding.
Now it keeps track of where in the string we are,
so it will only try to expand any index once.
Also improved handling of invalid keys like $. and $-.
Since this is not something we handle,
it is clearer to just leave it in place.